murtaoweb/requirements.txt
appdirs==1.4.3
astroid==2.3.3
beautifulsoup4==4.6.0
distlib==0.3.0
Django==2.0.3
django-contrib-comments==1.8.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-mptt==0.9.0
django-multiselectfield==0.1.12
django-simple-history==2.10.0
django-tagging==0.4.6
django-xmlrpc==0.1.8
easy-thumbnails==2.7
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mots-vides==2015.5.11
Pillow==7.1.2
pyparsing==2.2.0
pytz==2020.1
regex==2018.2.21
six==1.14.0
virtualenv==20.0.15
wrapt==1.11.2
xlwt==1.3.0